+++ packages/zk/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Breaking changes policy
 
+## 2.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @caatinga/core@2.4.1
+
 ## 2.4.0
 
 ### Minor Changes